fix up a few more tests
