+++ validator/rpc/auth_test.go
@@ -58,7 +58,7 @@ func TestServer_SignupAndLogin_RoundTrip(t *testing.T) {
 	require.NoError(t, err)
 
 	// Assert we stored the hashed password.
-	passwordHashExists := fileutil.FileExists(filepath.Join(defaultWalletPath, wallet.HashedPasswordFileName))
+	passwordHashExists := fileutil.FileExists(filepath.Join(defaultWalletPath, HashedRPCPassword))
 	assert.Equal(t, true, passwordHashExists)
 
 	// We attempt to create the wallet.
@@ -75,6 +75,50 @@ func TestServer_SignupAndLogin_RoundTrip(t *testing.T) {
 	// We assert we are able to login.
 	_, err = ss.Login(ctx, &pb.AuthRequest{
 		Password: strongPass,
+	})
+	require.NoError(t, err)
+}
+
+func TestServer_ChangePassword_Preconditions(t *testing.T) {
+	localWalletDir := setupWalletDir(t)
+	defaultWalletPath = localWalletDir
+	ctx := context.Background()
+	strongPass := "29384283xasjasd32%%&*@*#*"
+	ss := &Server{
+		walletDir: defaultWalletPath,
+	}
+	require.NoError(t, ss.SaveHashedPassword(strongPass))
+	_, err := ss.ChangePassword(ctx, &pb.ChangePasswordRequest{
+		CurrentPassword: strongPass,
+		Password:        "",
+	})
+	assert.ErrorContains(t, "Could not validate password input", err)
+	_, err = ss.ChangePassword(ctx, &pb.ChangePasswordRequest{
+		CurrentPassword:      strongPass,
+		Password:             "abc",
+		PasswordConfirmation: "def",
+	})
+	assert.ErrorContains(t, "does not match", err)
+}
+
+func TestServer_ChangePassword_OK(t *testing.T) {
+	localWalletDir := setupWalletDir(t)
+	defaultWalletPath = localWalletDir
+	ss := &Server{
+		walletDir: defaultWalletPath,
+	}
+	password := "Passw0rdz%%%%pass"
+	newPassword := "NewPassw0rdz%%%%pass"
+	ctx := context.Background()
+	require.NoError(t, ss.SaveHashedPassword(password))
+	_, err := ss.ChangePassword(ctx, &pb.ChangePasswordRequest{
+		CurrentPassword:      password,
+		Password:             newPassword,
+		PasswordConfirmation: newPassword,
+	})
+	require.NoError(t, err)
+	_, err = ss.Login(ctx, &pb.AuthRequest{
+		Password: newPassword,
 	})
 	require.NoError(t, err)
 }
