+++ validator/rpc/wallet_test.go
@@ -210,92 +210,6 @@ func TestServer_WalletConfig(t *testing.T) {
 	})
 }
 
-func TestServer_ChangePassword_Preconditions(t *testing.T) {
-	localWalletDir := setupWalletDir(t)
-	defaultWalletPath = localWalletDir
-	ctx := context.Background()
-	strongPass := "29384283xasjasd32%%&*@*#*"
-	ss := &Server{
-		walletDir: defaultWalletPath,
-	}
-	_, err := ss.ChangePassword(ctx, &pb.ChangePasswordRequest{
-		CurrentPassword: strongPass,
-		Password:        "",
-	})
-	assert.ErrorContains(t, noWalletMsg, err)
-	// We attempt to create the wallet.
-	w, err := accounts.CreateWalletWithKeymanager(ctx, &accounts.CreateWalletConfig{
-		WalletCfg: &wallet.Config{
-			WalletDir:      defaultWalletPath,
-			KeymanagerKind: keymanager.Derived,
-			WalletPassword: strongPass,
-		},
-		SkipMnemonicConfirm: true,
-	})
-	require.NoError(t, err)
-	km, err := w.InitializeKeymanager(ctx, true /* skip mnemonic confirm */)
-	require.NoError(t, err)
-	ss.wallet = w
-	ss.walletInitialized = true
-	ss.keymanager = km
-	_, err = ss.ChangePassword(ctx, &pb.ChangePasswordRequest{
-		CurrentPassword: strongPass,
-		Password:        "",
-	})
-	assert.ErrorContains(t, "Could not validate wallet password", err)
-	_, err = ss.ChangePassword(ctx, &pb.ChangePasswordRequest{
-		CurrentPassword:      strongPass,
-		Password:             "abc",
-		PasswordConfirmation: "def",
-	})
-	assert.ErrorContains(t, "does not match", err)
-}
-
-func TestServer_ChangePassword_ImportedKeymanager(t *testing.T) {
-	ss, _ := createImportedWalletWithAccounts(t, 1)
-	newPassword := "NewPassw0rdz%%%%pass"
-	_, err := ss.ChangePassword(context.Background(), &pb.ChangePasswordRequest{
-		CurrentPassword:      ss.wallet.Password(),
-		Password:             newPassword,
-		PasswordConfirmation: newPassword,
-	})
-	require.NoError(t, err)
-	assert.Equal(t, ss.wallet.Password(), newPassword)
-}
-
-func TestServer_ChangePassword_DerivedKeymanager(t *testing.T) {
-	localWalletDir := setupWalletDir(t)
-	defaultWalletPath = localWalletDir
-	ctx := context.Background()
-	strongPass := "29384283xasjasd32%%&*@*#*"
-	// We attempt to create the wallet.
-	w, err := accounts.CreateWalletWithKeymanager(ctx, &accounts.CreateWalletConfig{
-		WalletCfg: &wallet.Config{
-			WalletDir:      defaultWalletPath,
-			KeymanagerKind: keymanager.Derived,
-			WalletPassword: strongPass,
-		},
-		SkipMnemonicConfirm: true,
-	})
-	require.NoError(t, err)
-	km, err := w.InitializeKeymanager(ctx, true /* skip mnemonic confirm */)
-	require.NoError(t, err)
-	ss := &Server{
-		walletDir: defaultWalletPath,
-	}
-	ss.wallet = w
-	ss.walletInitialized = true
-	ss.keymanager = km
-	newPassword := "NewPassw0rdz%%%%pass"
-	_, err = ss.ChangePassword(ctx, &pb.ChangePasswordRequest{
-		CurrentPassword:      strongPass,
-		Password:             newPassword,
-		PasswordConfirmation: newPassword,
-	})
-	require.NoError(t, err)
-	assert.Equal(t, w.Password(), newPassword)
-}
-
 func TestServer_HasWallet(t *testing.T) {
 	localWalletDir := setupWalletDir(t)
 	defaultWalletPath = localWalletDir
